refactor(common) file 업로드및 삭제 관련 코드 공통 모듈 common로 이동
