+++ common/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     compileOnly 'org.springframework.boot:spring-boot-starter-security'
     compileOnly 'org.springframework.boot:spring-boot-starter-data-jpa'
     compileOnly 'org.springframework.boot:spring-boot-starter-kafka'
+    compileOnly 'org.springframework.boot:spring-boot-starter-web'
 
     //jackson-datatype-jsr310 의존성(Jackson이 직렬화)
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
